fix: server middleware разбить gzip по нескольким файлам
