+++ cmd/server/middleware/gzip.go
@@ -9,94 +9,28 @@ import (
 	"strings"
 
 	"github.com/thefrol/kysh-kysh-meow/internal/ololog"
-	"github.com/thefrol/kysh-kysh-meow/internal/slices"
 )
 
 // UnGZIP распаковывает запросы, закодированные при помощи GZIP, и пропускает все
 // остальное мимо ушей, отслеживает чтобы хотя бы один из заголовков Content-Encoding
 // содержал подстроку gzip
 func UnGZIP(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		// пропускаем, поскольку мы не обрабатываем такие сжималки
-		if !encoded(r, "gzip") {
-			next.ServeHTTP(w, r)
-			return
-		}
 
-		// в случае если перед нами gzip, заменяем исходное тело запроса на обертку с gzip
-		gz, err := gzip.NewReader(r.Body)
-		if err != nil {
-			ololog.Error().Str("location", "middleware/gzip").Strs("Content-Encoding", r.Header.Values("Content-Encoding")).Err(err).Msg("Cant unzip a request body")
-			http.Error(w, "cant unzip", http.StatusBadRequest)
-			//todo try recover and send data as is
-			return
-		}
-		defer r.Body.Close()
-		r.Body = gz //todo: make a pool of zgips and return only buffer, not gzipeers
-
-		next.ServeHTTP(w, r)
-
-	})
-}
-
-type gzipOptions struct {
-	MinLenght          int
-	contentTypes       []string
-	CompressionLevel   int
-	allowedStatusCodes []int
-}
-
-func (o gzipOptions) CheckContentTypes(h http.Header) bool {
-	for _, ct := range o.contentTypes {
-		for _, c := range h.Values("Content-Type") {
-			if strings.Contains(c, ct) {
-				return true
+	// Тут описываемся сама мидлварь, получившая opts
+	// в качестве настроек
+	return func(next http.Handler) http.Handler {
+		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			// Eсли клиент поддерживает gzip, то подменяем врайтер, не забыв его закрыть, и отправляем
+			// запрос дальше по цепочке
+			if acceptsEncoding(r, "gzip") {
+				cw := NewCompressedWriter(w, funcOpts...)
+				w = cw
+				defer cw.Close()
 			}
-		}
+			next.ServeHTTP(w, r)
+
+		})
 	}
-	return false
-}
-
-func (o gzipOptions) CheckStatusCode(code int) bool {
-	return slices.Contains[int](o.allowedStatusCodes, code)
-}
-
-type gzipFuncOpt func(*gzipOptions)
-
-func ContentTypes(ct ...string) gzipFuncOpt {
-	return func(opts *gzipOptions) {
-		opts.contentTypes = append(opts.contentTypes, ct...)
-	}
-}
-
-func MinLenght(len int) gzipFuncOpt {
-	return func(opts *gzipOptions) {
-		opts.MinLenght = len
-	}
-}
-
-func CompressionLevel(level int) gzipFuncOpt {
-	return func(opt *gzipOptions) {
-		opt.CompressionLevel = level
-	}
-
-}
-
-// GZIPBestCompression устанавливаем компрессию на уровень gzip.BestSpeed
-func GZIPBestSpeed(opt *gzipOptions) {
-	opt.CompressionLevel = gzip.BestSpeed
-}
-
-// GZIPBestCompression устанавливаем компрессию на уровень gzip.BestCompression
-func GZIPBestCompression(opt *gzipOptions) {
-	opt.CompressionLevel = gzip.BestCompression
-}
-
-func StatusCodes(codes ...int) gzipFuncOpt {
-	return func(opt *gzipOptions) {
-		opt.allowedStatusCodes = append(opt.allowedStatusCodes, codes...)
-	}
-
 }
 
 // CompressedWriter это обертка над ResponseWriter,
@@ -131,6 +65,26 @@ type CompressedWriter struct {
 	ignore  bool // true - не будем сжимать
 	checked bool // true- прошел все проверки по другим параментрам кроме минамальной длинны и собирается быть сжат
 
+}
+
+func NewCompressedWriter(originalWriter http.ResponseWriter, funcOpts ...gzipFuncOpt) *CompressedWriter {
+
+	// Создаем структуру настроек для GZIP,
+	// после применения к ней опцефункций,
+	// она будет передана в замыкание и будет уже работать с мидлварью
+	opts := gzipOptions{
+		CompressionLevel: gzip.BestCompression,
+	}
+	for _, f := range funcOpts {
+		f(&opts)
+	}
+	return &CompressedWriter{
+		opts:           opts,
+		originalWriter: originalWriter,
+		statusCode:     200, // сразу ставим StatusOK, потому что такой код у нас по умолчанию
+	}
+
+	//todo я думаю нам нужны какие-то настройки по умолчанию, заголовки и статус коды нарзерешенные
 }
 
 func (cw CompressedWriter) Header() http.Header {
@@ -274,56 +228,6 @@ func (cw *CompressedWriter) Close() {
 }
 
 var _ http.ResponseWriter = (*CompressedWriter)(nil)
-
-func GZIP(funcOpts ...gzipFuncOpt) func(http.Handler) http.Handler {
-
-	// Тут описываемся сама мидлварь, получившая opts
-	// в качестве настроек
-	return func(next http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			// Eсли клиент поддерживает gzip, то подменяем врайтер, не забыв его закрыть, и отправляем
-			// запрос дальше по цепочке
-			if acceptsEncoding(r, "gzip") {
-				cw := NewCompressedWriter(w, funcOpts...)
-				w = cw
-				defer cw.Close()
-			}
-			next.ServeHTTP(w, r)
-
-		})
-	}
-}
-
-func NewCompressedWriter(originalWriter http.ResponseWriter, funcOpts ...gzipFuncOpt) *CompressedWriter {
-
-	// Создаем структуру настроек для GZIP,
-	// после применения к ней опцефункций,
-	// она будет передана в замыкание и будет уже работать с мидлварью
-	opts := gzipOptions{
-		CompressionLevel: gzip.BestCompression,
-	}
-	for _, f := range funcOpts {
-		f(&opts)
-	}
-	return &CompressedWriter{
-		opts:           opts,
-		originalWriter: originalWriter,
-		statusCode:     200, // сразу ставим StatusOK, потому что такой код у нас по умолчанию
-	}
-}
-
-// encoded возвращает true, если запрос r закодирован
-// кодировщиком encoder. Например:
-//
-// if encoded(r,"gzip"){...}
-func encoded(r *http.Request, encoder string) bool {
-	for _, v := range r.Header.Values("Content-Encoding") {
-		if strings.Contains(v, encoder) {
-			return true
-		}
-	}
-	return false
-}
 
 // acceptsEncoding возвращает true, если на основании запроса r
 // можно сказать, клиент поддерживает кодирование в encoding
